refactor: Separando a lógica do histórico de posições em uma Store separada.
